게임 UI 추가
- 게임 화면에 Life, Score 표시 UI 작성

+++ Engine/Actor/Actor.cpp
@@ -15,7 +15,7 @@ namespace Wanted
 		: position(position), color(color)
 	{
 		// 문자열 복사.
-		int length = strlen(image);
+		int length = static_cast<int>(strlen(image));
 		for (int p = 0; p < length; ++p)
 		{
 			if (image[p] == '\n')
